weather_predict/requirements.txt
beautifulsoup4==4.13.4
itemadapter==0.3.0
joblib==1.4.2
lxml==5.3.0
matplotlib==3.3.4
numpy==1.22.4
pandas==1.1.4
PyQt5==5.15.11
PyQt5_sip==12.13.0
Requests==2.32.3
scikit_learn==1.6.1
Scrapy==2.12.0
tensorflow_gpu==2.6.0
torch==2.5.1
Twisted==24.11.0
xpinyin==0.7.6
